Fix bug where a school with no degree or major get parsed as having the same degree/major as the prior school in the resume
I encountered this issue when parsing Dan Bernier's resume. This was his "education array":

[{'School': 'Central Connecticut State University', 'Degree': 'BS', 'Major': 'Computer Science', 'Dates': {'from_month': '', 'from_year': '1997', 'to_month': '', 'to_year': '2006'}, 'Activities': ''}, {'School': 'Ruby Mendicant University', 'Degree': 'BS', 'Major': 'Computer Science', 'Dates': {'from_month': '', 'from_year': '2011', 'to_month': '', 'to_year': '2011'}, 'Activities': ''}]

However, Dan Bernier never specified a degree or major for Ruby Mendicant University. This is a bug in resume_parser.py, and it was due to the code reusing the previous degree and major from Central Connecticut State University. To deal with this issue, and to make the code easier to modify and understand in case future bugs are discovered, I refactored the code (creating multiple different methods and renaming variables).