Change algorithm for finding all instances of technological jargon
The previous algorithm would not be able to handle situations
where the "keyword" is two or more words (example: "Visual Basic").
We wrote a new algorithm using a regex, which was an incredibly
painful process (since the regex had to handle all kinds of
edge cases and unique situations),but at least it's now over and
done with.

Our final regex pattern finds all instances of a keyword that is
next to whitespaces/begin-of-string/end-of-string. We had
to do this because some keywords were single-letter languages, like C,
so if we don't check for spaces/begin-of-string/end-of-string, we will
have a lot of spurious keyword matches.

We also have to escape all special characters within the keyword in
question to prevent them from interfering with the final "regex".
For example, the "+"s in "C++" are special characters.