Add name and email to recruiter_json and candidate_json
We use a regex to extract out the name and email from a resume.
We also changed the format of the generated json as well. Now, the
final generated json contains the name, email, and the "counter"
(representing the frequency of tech keywords).

Example:
```
"candidate_json": "{\"counter\": {\"Ruby\": 15, \"JavaScript\": 2, \"Angular\": 3, \".NET\": 1, \"WordPress\": 3, \"C#\": 1, \"AWS\": 1, \"Capistrano\": 1, \"GitHub\": 1, \"React\": 1, \"Bootstrap\": 2, \"Ethereum\": 1, \"blockchain\": 1, \"C++\": 2, \"Rails\": 5, \"Sinatra\": 2, \"jQuery\": 2, \"Javascript\": 2, \"Twitter\": 2, \"RSpec\": 1}, \"name\": \"Tariq\\u00a0Ali\", \"email\": \"tra38@nau.edu\"}"
```

We updated the codebase to take into account the new format of the generated json.
We also wrote more automated tests to verify that we grabbing the
additional information (name and email) correctly.

The default regex for handling names fail on an edge case (all-caps names).
To handle this, I wrote an alternative regex. We compare the output
between the normal regex and the alternative regex, and choose the output
that have the most characters.

We also detected a edge case where the regex parser is unable to handle
correctly the existence of a "middle" name, but I cannot figure out
how to handle it effectively.