Change the proxy URL to just localhost
Using localhost/api generates the following url on
the backend "/api/api/profile", which is not a valid
endpoint.

CSRF protection is currently mandatory for all endpoints
unless we specifically opted out of them. We opt-out of
CSRF protection for "/api/profile", but not "/api/api/profile",
which led to a CSRF error being raised.